Rename base package and classes
