--- commonmark-ext-yaml-front-matter/src/main/java/org/commonmark/ext/yaml/YAMLFrontMatterBlock.java
@@ -1,6 +1,0 @@
-package org.commonmark.ext.yaml;
-
-import org.commonmark.node.CustomBlock;
-
-public class YAMLFrontMatterBlock extends CustomBlock {
-}
